Add enc_dec_fuzzer
This fuzzer takes input, compresses it, decompressed, and compresses
it again, and checks that the two compressed versions are equal.
This is a simple correctness fuzzer. Note that we can't just
compare the decompressed version to the original input because
decompression might produce a slightly different TTF file to the
input, even though they are functionally the exact same thing.